app/requirements.txt
#Django==1.7c1
https://www.djangoproject.com/download/1.7c1/tarball/
Markdown==2.4.1
argparse==1.2.1
dj-database-url==0.3.0
django-debug-toolbar==1.2.1
django-filter==0.7
django-redis==3.6.2
django-rq==0.6.2
django-social-auth==0.7.28
djangorestframework==2.3.14
fudge==1.0.3
httplib2==0.9
ipython==2.1.0
oauth2==1.5.211
psycopg2==2.5.3
python-openid==2.2.5
redis==2.10.1
requests==2.3.0
rq==0.4.6
simplejson==3.6.0
soundcloud==0.4.1
sqlparse==0.1.11
wsgiref==0.1.2
